tests better structure
